requirements.txt
streamlit==1.33.0
tensorflow==2.16.1
numpy
pillow
gdown